Reorganising components folders
