+++ src/components/WebcamCapture/index.jsx
@@ -1,7 +1,7 @@
 import React, { useRef, useState } from "react";
 import Webcam from "react-webcam";
-import ImagePreviewer from "./ImagePreviewer";
-import boardStyles from "./styles/Board.module.sass";
+import ImagePreviewer from "../ImagePreviewer";
+import boardStyles from "../Board/index.module.sass";
 import dashboardStyles from "@pages/Dashboard/index.module.sass";
 import { Camera } from "react-bootstrap-icons";
