+++ src/components/index.jsx
@@ -6,5 +6,8 @@ import MediaItem from "./MediaItem";
 import PopUpButtons from "./PopUpButtons";
 import Loading from "./Loading";
 import Marquee from "./Marquee/";
+import IsPrivate from "./IsPrivate/";
+import IsAnon from "./IsAnon/";
 
-export { Button, NavBar, Board, MediaForm, MediaItem, PopUpButtons, Loading, Marquee };
+export { Button, NavBar, Board, MediaForm, MediaItem, PopUpButtons, Loading, Marquee, IsPrivate, IsAnon
+ };
