+++ src/pages/Dashboard/index.jsx
@@ -11,8 +11,8 @@ import assetsService from '../../services/assets.service';
 import usersService from '../../services/users.service';
 import { Container, Row, Col } from 'react-bootstrap';
 import styles from './index.module.sass';
-import boardStyles from '@components/styles/Board.module.sass';
-import popUpButtonStyles from '@components/PopUpButtons/PopUpButtons.module.sass';
+import boardStyles from '@components/Board/index.module.sass';
+import popUpButtonStyles from '@components/PopUpButtons/index.module.sass';
 import { PlusLg } from 'react-bootstrap-icons';
 import classNames from 'classnames';
